chore: update async task error handling
