+++ src/speaches/realtime/utils.py
@@ -1,5 +1,10 @@
+import asyncio
+from contextvars import Context
+import logging
 import random
 import string
+
+logger = logging.getLogger(__name__)
 
 
 def generate_id_suffix() -> str:
@@ -24,3 +29,12 @@ def generate_session_id() -> str:
 
 def generate_call_id() -> str:
     return "call_" + generate_id_suffix()
+
+
+def task_done_callback(task: asyncio.Task, *, context: Context | None = None) -> None:  # noqa: ARG001
+    try:
+        task.result()
+    except asyncio.CancelledError:
+        logger.info(f"Task {task.get_name()} cancelled")
+    except BaseException:  # TODO: should this be `Exception` instead?
+        logger.exception(f"Task {task.get_name()} failed")
